buffer fix & cleanup
