+++ fmt/src/lib.rs
@@ -1,5 +1,6 @@
 #![doc = include_str!("../README.md")]
 
+mod buffer;
 pub mod chunk;
 mod comments;
 mod formatter;
