+++ fmt/testdata/IfStatement/fmt.sol
@@ -31,8 +31,8 @@ contract IfStatement {
         }
 
         /* comment1 */
-        if (
-            /* comment2 */ /* comment3 */ condition // comment4
+        if ( /* comment2 */ /* comment3 */
+            condition // comment4
         ) {
             // comment5
             execute();
@@ -43,14 +43,21 @@ contract IfStatement {
         } // comment7
         /* comment8 */
         /* comment9 */
-        else if (
-            /* comment10 */ anotherLongCondition // comment11
+        else if ( /* comment10 */
+            anotherLongCondition // comment11
         ) {
             /* comment12 */
             execute();
         } // comment13
         /* comment14 */
         else {} // comment15
+
+        if (
+            // comment16
+            condition /* comment17 */
+        ) {
+            execute();
+        }
 
         if (condition) {
             execute();
