+++ fmt/testdata/ModifierDefinition/fmt.sol
@@ -3,11 +3,7 @@ contract ModifierDefinitions {
     modifier noParams() {}
     modifier oneParam(uint256 a) {}
     modifier twoParams(uint256 a, uint256 b) {}
-    modifier threeParams(
-        uint256 a,
-        uint256 b,
-        uint256 c
-    ) {}
+    modifier threeParams(uint256 a, uint256 b, uint256 c) {}
     modifier fourParams(
         uint256 a,
         uint256 b,
